AI_Agent/ollama_backend/requirements.txt
# Ollama Backend Requirements
# Python 3.9+

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0  # Use newer version with pre-built wheels (no Rust needed)

# HTTP Client
httpx>=0.25.1

# Configuration
PyYAML>=6.0.1

# Utilities
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1

# Monitoring (optional)
prometheus-client>=0.19.0
structlog>=23.2.0
